+++ frontend/src/features/home/components/main/EmptyDirDisplay.tsx
@@ -5,7 +5,7 @@ import { VscFiles } from 'react-icons/vsc';
 import { UseMutationResult } from 'react-query';
 import { SendRequestMutationConfig } from '../../api/sendRequest';
 import { Uploads } from '../../api/upload';
-import { NewMenu } from '../misc/NewMenu';
+import { NewMenu } from '../misc/newmenu/NewMenu';
 
 type EmptyDirDisplayProps = {
   currentDir: string;
